working on styling

+++ clientsrc/src/components/Note.vue
@@ -1,12 +1,14 @@
 <template>
-<div class="note card bg-success col-12 mr-1 mb-2 p-2 border-0 rounded-lg">
+  <div class="note card bg-success col-12 mr-1 mb-2 p-2 border-0 rounded-lg">
     <div v-if="!editing">
       <div class="row">
         <div class="col-12">
           <div class="row">
-            <h5 class="col-6 p-0 ml-4  mt-1 mb-1"><strong>{{noteData.title}}</strong></h5>
-            <div class="col-1 ml-auto text-right">
-              <button class="btn btn-sm textColor" @click="editing = true">
+            <h5 class="col-6 p-0 mr-auto ml-4 mt-1 mb-1">
+              <strong>{{noteData.title}}</strong>
+            </h5>
+            <div class="col-1 ml-auto mr-4 text-right">
+              <!-- <button class="btn btn-sm textColor" @click="editing = true">
                 <svg
                   class="bi bi-pencil"
                   width="1em"
@@ -26,69 +28,70 @@
                     clip-rule="evenodd"
                   />
                 </svg>
-              </button>
-              <button class="btn btn-sm text-danger" @click="deleteNote()">
-                <svg
-                  class="bi bi-trash"
-                  width="1em"
-                  height="1em"
-                  viewBox="0 0 16 16"
-                  fill="currentColor"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M5.5 5.5A.5.5 0 016 6v6a.5.5 0 01-1 0V6a.5.5 0 01.5-.5zm2.5 0a.5.5 0 01.5.5v6a.5.5 0 01-1 0V6a.5.5 0 01.5-.5zm3 .5a.5.5 0 00-1 0v6a.5.5 0 001 0V6z"
-                  />
-                  <path
-                    fill-rule="evenodd"
-                    d="M14.5 3a1 1 0 01-1 1H13v9a2 2 0 01-2 2H5a2 2 0 01-2-2V4h-.5a1 1 0 01-1-1V2a1 1 0 011-1H6a1 1 0 011-1h2a1 1 0 011 1h3.5a1 1 0 011 1v1zM4.118 4L4 4.059V13a1 1 0 001 1h6a1 1 0 001-1V4.059L11.882 4H4.118zM2.5 3V2h11v1h-11z"
-                    clip-rule="evenodd"
-                  />
-                </svg>
-              </button>
-            </div>
+            </button>-->
+            <button class="btn btn-sm text-danger" @click="deleteNote()">
+              <svg
+                class="bi bi-trash"
+                width="1em"
+                height="1em"
+                viewBox="0 0 16 16"
+                fill="currentColor"
+                xmlns="http://www.w3.org/2000/svg"
+              >
+                <path
+                  d="M5.5 5.5A.5.5 0 016 6v6a.5.5 0 01-1 0V6a.5.5 0 01.5-.5zm2.5 0a.5.5 0 01.5.5v6a.5.5 0 01-1 0V6a.5.5 0 01.5-.5zm3 .5a.5.5 0 00-1 0v6a.5.5 0 001 0V6z"
+                />
+                <path
+                  fill-rule="evenodd"
+                  d="M14.5 3a1 1 0 01-1 1H13v9a2 2 0 01-2 2H5a2 2 0 01-2-2V4h-.5a1 1 0 01-1-1V2a1 1 0 011-1H6a1 1 0 011-1h2a1 1 0 011 1h3.5a1 1 0 011 1v1zM4.118 4L4 4.059V13a1 1 0 001 1h6a1 1 0 001-1V4.059L11.882 4H4.118zM2.5 3V2h11v1h-11z"
+                  clip-rule="evenodd"
+                />
+              </svg>
+            </button>
           </div>
         </div>
       </div>
     </div>
-    <div v-else>
+  </div>
+  <!-- <div v-else>
       <form @submit.prevent="editNote()">
         <input type="text" v-model="noteData.title" />
         <button class="btn text-warning">submit</button>
       </form>
-    </div>
+  </div>-->
   </div>
 </template>
 
 
 <script>
 export default {
-  name: 'note',
-  props:["noteData"],
-  data(){
+  name: "note",
+  props: ["noteData"],
+  data() {
     return {
       editing: false
-    }
+    };
   },
-  computed:{
+  computed: {
     profile() {
       return this.$store.state.user;
-    },
+    }
   },
-  methods:{
-      deleteNote() {
+  methods: {
+    deleteNote() {
+    if (confirm("are you sure you want to close this ticket?")) {
       this.$store.dispatch("deleteNote", this.noteData);
+    }
     },
     editNote() {
       this.$store.dispatch("editNote", this.noteData);
       this.editing = false;
     }
   },
-  components:{}
-}
+  components: {}
+};
 </script>
 
 
 <style scoped>
-
 </style>
